pyproject.toml
[tool.poetry]
name = "mod-hyde"
version = "0.1.0"
description = ""
authors = ["Athe-kunal <athekunal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = {extras = ["torch"], version = "^4.42.3"}
langchain-text-splitters = "^0.2.2"
python-dotenv = "^1.0.1"
wandb = "^0.17.4"
datasets = "^2.20.0"
accelerate = "^0.32.0"
dspy-ai = "^2.4.10"
chromadb = "^0.5.3"
fire = "^0.6.0"
moviepy = "^1.0.3"
unstructured = "0.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
